updated matlab code

+++ wave.java
@@ -21,17 +21,14 @@ class wave
 	 for (int i = 0; i < 65; i++)
 		zernikes[i] = Float.parseFloat(args[i+1]); 
 	 MWNumericArray coeffs = new MWNumericArray(zernikes, MWClassID.DOUBLE);
-//	 float[] coeffs = new float[65];
-//	 for (int i = 0; i < 65; i++)
-//		coeffs[i] = Float.parseFloat(args[i+1]);
-	 int pupilfit = Integer.parseInt(args[66]);
-	 int pupilcalc = Integer.parseInt(args[67]);
-	 int defocus = Integer.parseInt(args[68]);
-	 float wavelength = Float.parseFloat(args[69]);
-	 int pixels = Integer.parseInt(args[70]);
-	 int pupilfieldsize = Integer.parseInt(args[71]);
-	 int lettersize = Integer.parseInt(args[72]);
-	 
+         float pupilfit = Float.parseFloat(args[66]);
+         float pupilcalc = Float.parseFloat(args[67]);
+         float defocus = Float.parseFloat(args[68]);
+         float wavelength = Float.parseFloat(args[69]);
+         float pixels = Float.parseFloat(args[70]);
+         float pupilfieldsize = Float.parseFloat(args[71]);
+         float lettersize = Float.parseFloat(args[72]); 
+
 	 int WF = Integer.parseInt(args[73]);
 	 int PSF = Integer.parseInt(args[74]);
 	 int MTF = Integer.parseInt(args[75]);
